tls-gatherer/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
astroid==2.3.3
    # via pylint
attrs==19.3.0
    # via pytest
certifi==2020.4.5.1
    # via requests
chardet==3.0.4
    # via requests
coverage==5.0.4
    # via pytest-cov
docker==4.2.0
    # via -r requirements.in
entrypoints==0.3
    # via flake8
flake8==3.7.9
    # via pytest-flake8
idna==2.9
    # via requests
importlib-metadata==1.5.0
    # via
    #   pluggy
    #   pytest
isort==4.3.21
    # via pylint
lazy-object-proxy==1.4.3
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
more-itertools==8.2.0
    # via pytest
mysql-connector==2.2.9
    # via -r requirements.in
packaging==20.3
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.8.1
    # via pytest
pycodestyle==2.5.0
    # via flake8
pyflakes==2.1.1
    # via flake8
pylint==2.4.4
    # via pytest-pylint
pyparsing==2.4.6
    # via packaging
pytest-asyncio==0.10.0
    # via -r requirements.in
pytest-cov==2.8.1
    # via -r requirements.in
pytest-flake8==1.0.4
    # via -r requirements.in
pytest-mock==2.0.0
    # via -r requirements.in
pytest-pylint==0.15.1
    # via -r requirements.in
pytest==5.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-flake8
    #   pytest-mock
    #   pytest-pylint
redis==3.5.3
    # via -r requirements.in
requests==2.23.0
    # via docker
six==1.14.0
    # via
    #   astroid
    #   docker
    #   packaging
    #   websocket-client
typed-ast==1.4.1
    # via astroid
urllib3==1.25.8
    # via requests
wcwidth==0.1.9
    # via pytest
websocket-client==0.57.0
    # via docker
wrapt==1.11.2
    # via astroid
zipp==3.1.0
    # via importlib-metadata
